Refactor CLI commands into plugins
