+++ betty/extension/nginx/_cli.py
@@ -6,7 +6,7 @@ import asyncio
 
 import click
 
-from betty.cli import pass_project, command
+from betty.cli.commands import command, pass_project
 from betty.extension.nginx import serve
 from betty.project import Project
 
@@ -14,7 +14,7 @@ from betty.project import Project
 @click.command(help="Serve a generated site with nginx in a Docker container.")
 @pass_project
 @command
-async def _serve(project: Project) -> None:
+async def serve_nginx_docker(project: Project) -> None:
     async with serve.DockerizedNginxServer(project) as server:
         await server.show()
         while True:
