+++ betty/tests/cli/test___init__.py
@@ -2,28 +2,33 @@ import logging
 from asyncio import to_thread
 from collections.abc import AsyncIterator
 from typing import Any
-from unittest.mock import AsyncMock
 
 import click
 import pytest
 from _pytest.logging import LogCaptureFixture
-from aiofiles.os import makedirs
+from click.testing import CliRunner, Result
+from pytest_mock import MockerFixture
+
 from betty.app import App
-from betty.cli import main, command, catch_exceptions
+from betty.cli import main, catch_exceptions
+from betty.cli.commands import command, Command
 from betty.config import write_configuration_file
 from betty.error import UserFacingError
 from betty.locale import DEFAULT_LOCALIZER
 from betty.locale.localizable import plain
+from betty.plugin.static import StaticPluginRepository
 from betty.project import Project
 from betty.serve import Server, ProjectServer
-from click.testing import CliRunner, Result
-from pytest_mock import MockerFixture
 
 
 @click.command(name="no-op")
 @command
 async def _no_op_command() -> None:
     pass
+
+
+class _NoOpCommand(Command):
+    _click_command = _no_op_command
 
 
 def run(
@@ -119,91 +124,9 @@ class NoOpProjectServer(NoOpServer, ProjectServer):
     pass
 
 
-class TestDemo:
-    async def test(self, mocker: MockerFixture) -> None:
-        mocker.patch("asyncio.sleep", side_effect=KeyboardInterrupt)
-        mocker.patch("betty.extension.demo.DemoServer", new=NoOpServer)
-
-        await to_thread(run, "demo")
-
-
-class TestDocs:
-    async def test(self, mocker: MockerFixture, new_temporary_app: App) -> None:
-        mocker.patch("asyncio.sleep", side_effect=KeyboardInterrupt)
-        mocker.patch("betty.documentation.DocumentationServer", new=NoOpServer)
-
-        await to_thread(run, "docs")
-
-
-class TestGenerate:
-    async def test(self, mocker: MockerFixture, new_temporary_app: App) -> None:
-        m_generate = mocker.patch("betty.generate.generate", new_callable=AsyncMock)
-        m_load = mocker.patch("betty.load.load", new_callable=AsyncMock)
-
-        async with Project.new_temporary(new_temporary_app) as project:
-            await write_configuration_file(
-                project.configuration, project.configuration.configuration_file_path
-            )
-            await to_thread(
-                run,
-                "generate",
-                "-c",
-                str(project.configuration.configuration_file_path),
-            )
-
-            m_load.assert_called_once()
-            await_args = m_load.await_args
-            assert await_args is not None
-            load_args, _ = await_args
-            assert (
-                load_args[0].configuration.configuration_file_path
-                == project.configuration.configuration_file_path
-            )
-
-            m_generate.assert_called_once()
-            generate_args, _ = m_generate.call_args
-            assert (
-                generate_args[0].configuration.configuration_file_path
-                == project.configuration.configuration_file_path
-            )
-
-
-class TestServe:
-    async def test(self, mocker: MockerFixture, new_temporary_app: App) -> None:
-        mocker.patch("asyncio.sleep", side_effect=KeyboardInterrupt)
-        mocker.patch("betty.serve.BuiltinProjectServer", new=NoOpProjectServer)
-        async with Project.new_temporary(new_temporary_app) as project:
-            await write_configuration_file(
-                project.configuration, project.configuration.configuration_file_path
-            )
-            await makedirs(project.configuration.www_directory_path)
-
-            await to_thread(
-                run, "serve", "-c", str(project.configuration.configuration_file_path)
-            )
-
-
 class TestUnknownCommand:
     async def test(self) -> None:
         await to_thread(run, "unknown-command", expected_exit_code=2)
-
-
-class TestInitTranslation:
-    async def test(self, mocker: MockerFixture) -> None:
-        locale = "nl-NL"
-        m_init_translation = mocker.patch("betty.locale.init_translation")
-        await to_thread(run, "init-translation", locale)
-        m_init_translation.assert_awaited_once_with(locale)
-
-    async def test_without_locale_arg(self) -> None:
-        await to_thread(run, "init-translation", expected_exit_code=2)
-
-
-class TestUpdateTranslations:
-    async def test(self, mocker: MockerFixture) -> None:
-        m_update_translations = mocker.patch("betty.locale.update_translations")
-        await to_thread(run, "update-translations")
-        m_update_translations.assert_awaited_once()
 
 
 class TestVerbosity:
@@ -218,9 +141,10 @@ class TestVerbosity:
     async def test(
         self, mocker: MockerFixture, new_temporary_app: App, verbosity: str
     ) -> None:
+        command_repository = StaticPluginRepository(_NoOpCommand)
         mocker.patch(
-            "betty.cli._discover.discover_commands",
-            return_value={"no-op": _no_op_command},
+            "betty.cli.commands.COMMAND_REPOSITORY",
+            new=command_repository,
         )
         async with Project.new_temporary(new_temporary_app) as project:
             await write_configuration_file(
